move 'case search' tests to the case_search app
