+++ corehq/apps/es/tests/test_case_search_es.py
@@ -1,31 +1,13 @@
 import uuid
-from collections import Counter
 from datetime import date, datetime
+from unittest.mock import MagicMock, patch
 
 from django.test import TestCase
 from django.test.testcases import SimpleTestCase
 
-from unittest.mock import MagicMock, patch
-
-from pillowtop.es_utils import initialize_index_and_mapping
-
-from corehq.apps.app_manager.models import (
-    Application,
-    CaseSearchProperty,
-    DetailColumn,
-    Module,
-)
 from corehq.apps.case_search.const import IS_RELATED_CASE, RELEVANCE_SCORE
 from corehq.apps.case_search.models import (
     CaseSearchConfig,
-    FuzzyProperties,
-    IgnorePatterns,
-)
-from corehq.apps.case_search.utils import (
-    _QueryHelper,
-    get_related_case_relationships,
-    get_related_case_results,
-    get_related_cases,
 )
 from corehq.apps.es import queries
 from corehq.apps.es.case_search import (
@@ -36,7 +18,7 @@ from corehq.apps.es.case_search import (
     flatten_result,
     wrap_case_search_hit,
 )
-from corehq.apps.es.tests.utils import ElasticTestMixin, es_test, get_case_search_query
+from corehq.apps.es.tests.utils import ElasticTestMixin, es_test
 from corehq.elastic import SIZE_LIMIT, get_es_new
 from corehq.form_processor.models import CommCareCaseIndex
 from corehq.form_processor.tests.utils import FormProcessorTestUtils
@@ -47,6 +29,7 @@ from corehq.pillows.mappings.case_search_mapping import (
 )
 from corehq.util.elastic import ensure_index_deleted
 from corehq.util.test_utils import create_and_save_a_case
+from pillowtop.es_utils import initialize_index_and_mapping
 
 
 @es_test
@@ -372,12 +355,12 @@ class TestCaseSearchHitConversions(SimpleTestCase):
 
 
 @es_test
-class TestCaseSearchLookups(TestCase):
+class BaseCaseSearchTest(TestCase):
 
     def setUp(self):
         self.domain = 'case_search_es'
         self.case_type = 'person'
-        super(TestCaseSearchLookups, self).setUp()
+        super(BaseCaseSearchTest, self).setUp()
         FormProcessorTestUtils.delete_all_cases()
         self.elasticsearch = get_es_new()
         ensure_index_deleted(CASE_SEARCH_INDEX)
@@ -387,7 +370,7 @@ class TestCaseSearchLookups(TestCase):
 
     def tearDown(self):
         ensure_index_deleted(CASE_SEARCH_INDEX)
-        super(TestCaseSearchLookups, self).tearDown()
+        super(BaseCaseSearchTest, self).tearDown()
 
     def _make_case(self, domain, case_properties, index=None):
         # make a case
@@ -427,6 +410,8 @@ class TestCaseSearchLookups(TestCase):
         self.addCleanup(config.delete)
         return config
 
+
+class TestCaseSearchLookups(BaseCaseSearchTest):
     def test_simple_case_property_query(self):
         self._assert_query_runs_correctly(
             self.domain,
@@ -563,254 +548,3 @@ class TestCaseSearchLookups(TestCase):
             "dob >= '2020-03-02' and dob <= '2020-03-03'",
             ['c2', 'c3']
         )
-
-    def test_date_range_criteria(self):
-        self._create_case_search_config()
-        self._assert_query_runs_correctly(
-            self.domain,
-            [
-                {'_id': 'c1', 'dob': date(2020, 3, 1)},
-                {'_id': 'c2', 'dob': date(2020, 3, 2)},
-                {'_id': 'c3', 'dob': date(2020, 3, 3)},
-                {'_id': 'c4', 'dob': date(2020, 3, 4)},
-            ],
-            get_case_search_query(
-                self.domain,
-                [self.case_type],
-                {'dob': '__range__2020-03-02__2020-03-03'},
-            ),
-            None,
-            ['c2', 'c3']
-        )
-
-    def test_get_related_case_relationships(self):
-        app = Application.new_app(self.domain, "Case Search App")
-        module = app.add_module(Module.new_module("Search Module", "en"))
-        module.case_type = self.case_type
-        detail = module.case_details.short
-        detail.columns.extend([
-            DetailColumn(header={"en": "x"}, model="case", field="x", format="plain"),
-            DetailColumn(header={"en": "y"}, model="case", field="parent/parent/y", format="plain"),
-            DetailColumn(header={"en": "z"}, model="case", field="host/z", format="plain"),
-        ])
-        module.search_config.properties = [CaseSearchProperty(
-            name="texture",
-            label={"en": "Texture"},
-        )]
-
-        module = app.add_module(Module.new_module("Non-Search Module", "en"))
-        module.case_type = self.case_type
-        detail = module.case_details.short
-        detail.columns.append(
-            DetailColumn(header={"en": "zz"}, model="case", field="parent/zz", format="plain"),
-        )
-
-        self.assertEqual(get_related_case_relationships(app, self.case_type), {"parent/parent", "host"})
-        self.assertEqual(get_related_case_relationships(app, "monster"), set())
-
-    def test_get_related_case_results(self):
-        # Note that cases must be defined before other cases can reference them
-        cases = [
-            {'_id': 'c1', 'case_type': 'monster', 'description': 'grandparent of first person'},
-            {'_id': 'c2', 'case_type': 'monster', 'description': 'parent of first person', 'index': {
-                'parent': ('monster', 'c1')
-            }},
-            {'_id': 'c3', 'case_type': 'monster', 'description': 'parent of host'},
-            {'_id': 'c4', 'case_type': 'monster', 'description': 'host of second person', 'index': {
-                'parent': ('monster', 'c3')
-            }},
-            {'_id': 'c5', 'description': 'first person', 'index': {
-                'parent': ('monster', 'c2')
-            }},
-            {'_id': 'c6', 'description': 'second person', 'index': {
-                'host': ('monster', 'c4')
-            }},
-        ]
-        self._bootstrap_cases_in_es_for_domain(self.domain, cases)
-
-        hits = CaseSearchES().domain(self.domain).case_type(self.case_type).run().hits
-        cases = [wrap_case_search_hit(result) for result in hits]
-        self.assertEqual({case.case_id for case in cases}, {'c5', 'c6'})
-
-        self._assert_related_case_ids(cases, set(), set())
-        self._assert_related_case_ids(cases, {"parent"}, {"c2"})
-        self._assert_related_case_ids(cases, {"host"}, {"c4"})
-        self._assert_related_case_ids(cases, {"parent/parent"}, {"c1"})
-        self._assert_related_case_ids(cases, {"host/parent"}, {"c3"})
-        self._assert_related_case_ids(cases, {"host", "parent"}, {"c2", "c4"})
-        self._assert_related_case_ids(cases, {"host", "parent/parent"}, {"c4", "c1"})
-
-    def test_get_related_case_results_duplicates(self):
-        """Test that `get_related_cases` does not include any cases that are in the initial
-        set or are duplicates of others already found."""
-
-        # d1 :> c2 > c1 > a1
-        # d1 > c1
-        # Search for case type 'c'
-        # - initial results c1, c2
-        # - related lookups (parent, parent/parent) yield a1, c1, a1
-        # - child lookups yield c2, d1
-        # - (future) extension lookups yield d1
-        cases = [
-            {'_id': 'a1', 'case_type': 'a'},
-            {'_id': 'c1', 'case_type': 'c', 'index': {
-                'parent': ('a', 'a1'),
-            }},
-            {'_id': 'c2', 'case_type': 'c', 'index': {
-                'parent': ('c', 'c1'),
-            }},
-            {'_id': 'd1', 'case_type': 'd', 'index': {
-                'parent': ('c', 'c1'),
-                'host': ('c', 'c2'),
-            }},
-        ]
-        self._bootstrap_cases_in_es_for_domain(self.domain, cases)
-
-        hits = CaseSearchES().domain(self.domain).case_type("c").run().hits
-        cases = [wrap_case_search_hit(result) for result in hits]
-        self.assertEqual({case.case_id for case in cases}, {'c1', 'c2'})
-
-        with patch("corehq.apps.case_search.utils.get_related_case_relationships",
-                   return_value={"parent", "parent/parent"}), \
-             patch("corehq.apps.case_search.utils.get_child_case_types", return_value={"c", "d"}), \
-             patch("corehq.apps.case_search.utils.get_app_cached"):
-            cases = get_related_cases(_QueryHelper(self.domain), None, {"c"}, cases, None)
-
-        case_ids = Counter([case.case_id for case in cases])
-        self.assertEqual(set(case_ids), {"a1", "d1"})  # c1, c2 excluded since they are in the initial list
-        self.assertEqual(max(case_ids.values()), 1, case_ids)  # no duplicates
-
-    def test_get_related_case_results_expanded_results(self):
-        """Test that `get_related_cases` includes related cases for cases loaded
-        via the 'custom_related_case_property'."""
-
-        # Search for case type 'a'
-        # - initial results a[1-4]
-        # - expanded lookup yields b1
-        # - related lookups (parent) yield p1 (b1 -> p1)
-        # - child lookups yield c1 (b1 <- c1)
-        cases = [
-            {'_id': 'p1', 'case_type': 'p'},
-            {'_id': 'a1', 'case_type': 'a'},
-            {'_id': 'a2', 'case_type': 'a', 'custom_related_case_id': 'b1'},
-            {'_id': 'a3', 'case_type': 'a', 'custom_related_case_id': 'b1'},
-            {'_id': 'a4', 'case_type': 'a', 'custom_related_case_id': ''},
-            {'_id': 'b1', 'case_type': 'b', 'index': {
-                'parent': ('p', 'p1'),
-            }},
-            {'_id': 'c1', 'case_type': 'c', 'index': {
-                'parent': ('b', 'b1'),
-            }},
-        ]
-        self._bootstrap_cases_in_es_for_domain(self.domain, cases)
-
-        hits = CaseSearchES().domain(self.domain).case_type("a").run().hits
-        cases = [wrap_case_search_hit(result) for result in hits]
-        self.assertEqual({case.case_id for case in cases}, {'a1', 'a2', 'a3', 'a4'})
-
-        with patch("corehq.apps.case_search.utils.get_related_case_relationships",
-                   return_value={"parent"}), \
-             patch("corehq.apps.case_search.utils.get_child_case_types", return_value={"c"}), \
-             patch("corehq.apps.case_search.utils.get_app_cached"):
-            cases = get_related_cases(_QueryHelper(self.domain), None, {"a"}, cases, 'custom_related_case_id')
-
-        case_ids = Counter([case.case_id for case in cases])
-        self.assertEqual(set(case_ids), {"b1", "p1", "c1"})
-        self.assertEqual(max(case_ids.values()), 1, case_ids)  # no duplicates
-
-    def _assert_related_case_ids(self, cases, paths, ids):
-        results = get_related_case_results(_QueryHelper(self.domain), cases, paths)
-        result_ids = Counter([result.case_id for result in results])
-        self.assertEqual(ids, set(result_ids))
-        if result_ids:
-            self.assertEqual(1, max(result_ids.values()), result_ids)  # no duplicates
-
-    def test_fuzzy_properties(self):
-        cases = [
-            {'_id': 'c1', 'case_type': 'song', 'description': 'New York'},
-            {'_id': 'c2', 'case_type': 'song', 'description': 'Neu York'},
-            {'_id': 'c3', 'case_type': 'show', 'description': 'Boston'},
-        ]
-        config = self._create_case_search_config()
-        fuzzy_properties = FuzzyProperties.objects.create(
-            domain=self.domain,
-            case_type='song',
-            properties=['description'],
-        )
-        config.fuzzy_properties.add(fuzzy_properties)
-        self.addCleanup(fuzzy_properties.delete)
-        self._assert_query_runs_correctly(
-            self.domain,
-            cases,
-            get_case_search_query(self.domain, ['song', 'show'], {'description': 'New York'}),
-            None,
-            ['c1', 'c2']
-        )
-
-    def test_ignore_patterns(self):
-        cases = [
-            {'_id': 'c1', 'case_type': 'person', 'phone_number': '8675309'},
-            {'_id': 'c2', 'case_type': 'person', 'phone_number': '9045555555'},
-        ]
-        config = self._create_case_search_config()
-        pattern = IgnorePatterns.objects.create(
-            domain=self.domain, case_type='person', case_property='phone_number', regex="+1")
-        config.ignore_patterns.add(pattern)
-        self.addCleanup(pattern.delete)
-        self._assert_query_runs_correctly(
-            self.domain,
-            cases,
-            get_case_search_query(self.domain, ['person'], {'phone_number': '+18675309'}),
-            None,
-            ['c1']
-        )
-
-    def test_multiple_case_types(self):
-        cases = [
-            {'_id': 'c1', 'case_type': 'song', 'description': 'New York'},
-            {'_id': 'c2', 'case_type': 'song', 'description': 'Another Song'},
-            {'_id': 'c3', 'case_type': 'show', 'description': 'New York'},
-            {'_id': 'c4', 'case_type': 'show', 'description': 'Boston'},
-        ]
-        self._create_case_search_config()
-        self._assert_query_runs_correctly(
-            self.domain,
-            cases,
-            get_case_search_query(self.domain, ['show', 'song'], {'description': 'New York'}),
-            None,
-            ['c1', 'c3']
-        )
-
-    def test_blank_case_search(self):
-        # foo = '' should match all cases where foo is empty or absent
-        self._create_case_search_config()
-        self._bootstrap_cases_in_es_for_domain(self.domain, [
-            {'_id': 'c1', 'foo': 'redbeard'},
-            {'_id': 'c2', 'foo': 'blackbeard'},
-            {'_id': 'c3', 'foo': ''},
-            {'_id': 'c4'},
-        ])
-        for criteria, expected in [
-            ({'foo': ''}, ['c3', 'c4']),
-            ({'foo': ['', 'blackbeard']}, ['c2', 'c3', 'c4']),
-        ]:
-            actual = get_case_search_query(self.domain, [self.case_type], criteria).get_ids()
-            msg = f"{criteria} yielded {actual}, not {expected}"
-            self.assertItemsEqual(actual, expected, msg=msg)
-
-    def test_blank_case_search_parent(self):
-        self._create_case_search_config()
-        self._bootstrap_cases_in_es_for_domain(self.domain, [
-            {'_id': 'c1', 'foo': 'redbeard'},
-            {'_id': 'c2', 'case_type': 'child', 'index': {'parent': (self.case_type, 'c1')}},
-            {'_id': 'c3', 'foo': 'blackbeard'},
-            {'_id': 'c4', 'case_type': 'child', 'index': {'parent': (self.case_type, 'c3')}},
-            {'_id': 'c5', 'foo': ''},
-            {'_id': 'c6', 'case_type': 'child', 'index': {'parent': (self.case_type, 'c5')}},
-            {'_id': 'c7'},
-            {'_id': 'c8', 'case_type': 'child', 'index': {'parent': (self.case_type, 'c7')}},
-        ])
-        actual = get_case_search_query(self.domain, ['child'], {
-            'parent/foo': ['', 'blackbeard'],
-        }).get_ids()
-        self.assertItemsEqual(actual, ['c4', 'c6', 'c8'])
